@@ -1,11 +1,14 @@
 {
 	"name": "adrenalin",
+	"type": "module",
 	"main": "main.js",
-	"type": "commonjs",
 	"version": "1.1.1",
 	"author": "Armageddion",
 	"description": "Adrenalin CRM application",
 	"license": "MIT",
+	"repository": {
+		"url": "https://github.com/armageddon/adrenalin"
+	},
 	"scripts": {
 		"start": "electron-forge start",
 		"css": "npx @tailwindcss/cli -i ./src/assets/styles.css -o ./public/styles.css",
@@ -25,10 +28,11 @@
 		"test": "npm run test:unit && npm run test:integration && npm run test:e2e",
 		"electron": "electron .",
 		"electron:dev": "NODE_ENV=development electron .",
-		"build:electron": "npx electron-builder",
 		"package": "electron-forge package --platform=win32 --arch=ia32",
 		"publish": "electron-forge publish",
-		"make": "electron-forge make"
+		"make": "electron-forge make",
+		"make:zip": "electron-forge make --targets @electron-forge/maker-zip",
+		"make:appimage": "electron-forge make --targets @reforged/maker-appimage"
 	},
 	"dependencies": {
 		"@hono/node-server": "^1.19.5",
@@ -36,12 +40,14 @@
 		"alpinejs": "^3.15.0",
 		"better-sqlite3": "^12.4.1",
 		"electron-squirrel-startup": "^1.0.1",
+		"electron-store": "^11.0.2",
 		"hono": "^4.9.9",
 		"htmx.org": "^2.0.7",
-		"signature_pad": "^5.1.1"
+		"signature_pad": "^5.1.1",
+		"update-electron-app": "^3.1.1"
 	},
 	"devDependencies": {
-		"@biomejs/biome": "^2.2.6",
+		"@biomejs/biome": "^2.3.6",
 		"@electron-forge/cli": "^7.10.2",
 		"@electron-forge/maker-deb": "^7.10.2",
 		"@electron-forge/maker-rpm": "^7.10.2",
@@ -51,6 +57,7 @@
 		"@electron-forge/plugin-fuses": "^7.10.2",
 		"@electron-forge/publisher-github": "^7.10.2",
 		"@electron/fuses": "^1.8.0",
+		"@reforged/maker-appimage": "^5.1.0",
 		"@types/better-sqlite3": "^7.6.13",
 		"@types/jsdom": "^27.0.0",
 		"@types/node": "^24.5.2",
@@ -58,7 +65,7 @@
 		"@vitest/coverage-v8": "^3.2.4",
 		"@vitest/ui": "^3.2.4",
 		"cypress": "^15.4.0",
-		"electron": "33.0.0",
+		"electron": "35.0.0",
 		"jsdom": "^27.0.0",
 		"nodemon": "^3.1.10",
 		"patch-package": "^8.0.1",
