@@ -1,5 +1,5 @@
-const { FusesPlugin } = require('@electron-forge/plugin-fuses')
-const { FuseV1Options, FuseVersion } = require('@electron/fuses')
+import { FusesPlugin } from '@electron-forge/plugin-fuses'
+import { FuseV1Options, FuseVersion } from '@electron/fuses'
 
 module.exports = {
 	packagerConfig: {
@@ -26,10 +26,6 @@ module.exports = {
 		{
 			name: '@electron-forge/maker-zip',
 			platforms: ['darwin'],
-		},
-		// Linux (Ubuntu/Debian) - generated by Linux job
-		{
-			name: '@electron-forge/maker-deb',
 			config: {
 				options: {
 					maintainer: 'Armageddion',
@@ -37,13 +33,27 @@ module.exports = {
 				},
 			},
 		},
-		// Linux (Fedora/RedHat) - generated by Linux job
+		// Linux (Zip) - simple archive for Linux
 		{
-			name: '@electron-forge/maker-rpm',
+			name: '@electron-forge/maker-zip',
+			platforms: ['linux'],
 			config: {
 				options: {
 					maintainer: 'Armageddion',
 					homepage: 'https://github.com/armageddion/adrenalin',
+				},
+			},
+		},
+		// Linux (AppImage) - works on any Linux distribution
+		{
+			name: '@reforged/maker-appimage',
+			platforms: ['linux'],
+			config: {
+				options: {
+					maintainer: 'Armageddion',
+					homepage: 'https://github.com/armageddion/adrenalin',
+					icon: 'src/assets/logo.svg',
+					categories: ['Office'],
 				},
 			},
 		},
